Moved common code from application api route classes to base class
